# output some perf timing to stderr