--- a/bin/eslint.js
+++ b/bin/eslint.js
@@ -9,7 +9,7 @@ function runWithTiming(name, fn) {
   var start = new Date(),
       rv = fn(),
       duration = (new Date() - start) / 1000;
-  console.error("eslint.timing." + name + ": " + duration + "s")
+  console.error("eslint.timing." + name + ": " + duration + "s");
   return rv;
 }