@GordonDiggs 2015-08-21
Yep generally the concept looks good! What do you think of using node's `console.time`? https://nodejs.org/api/console.html#console_console_time_label