@GordonDiggs 2015-08-21
Not sure if it's easy/possible to direct that output to stderr though