@wfleming 2015-08-21
I think `console.time` outputs to stdout, not stderr. Couldn't see a simple way to redirect it, so i just wrote it myself.